+++ bitset_test.go
@@ -12,7 +12,7 @@ import (
 )
 
 func TestBitSet(t *testing.T) {
-	contains := func(i uint, values []uint) bool {
+	contains := func(i uint64, values []uint64) bool {
 		for _, v := range values {
 			if i == v {
 				return true
@@ -23,16 +23,16 @@ func TestBitSet(t *testing.T) {
 
 	seed := time.Now().UnixNano()
 	r := rand.New(rand.NewSource(seed))
-	size := uint(r.Intn(12345))
+	size := uint64(r.Intn(12345))
 	s := newBitset(size)
-	values := make([]uint, 0)
-	for i, count := uint(0), uint(r.Intn(100)); i < count; i++ {
-		values = append(values, uint(r.Intn(int(size))))
+	values := make([]uint64, 0)
+	for i, count := uint64(0), uint64(r.Intn(100)); i < count; i++ {
+		values = append(values, uint64(r.Intn(int(size))))
 	}
 	for _, v := range values {
 		s.set(v)
 	}
-	for i := uint(0); i < size; i++ {
+	for i := uint64(0); i < size; i++ {
 		if contains(i, values) {
 			if !s.isSet(i) {
 				t.Errorf("expected value %v is not set (seed %v)", i, seed)
